Organize sources more (read: at all)
Sources are now in one of three places:
* `src/cmds/` for commands
* `src/utils/` for utility functions
* `src/` for core bot stuff like the main script and deploy script
