+++ src/cmds/auto-publish.ts
@@ -1,4 +1,4 @@
-import { ComplexCommand } from './commands';
+import { ComplexCommand } from '../commands';
 import {
   Client,
   ColorResolvable,
@@ -11,8 +11,8 @@ import {
   Permissions,
   TextChannel
 } from 'discord.js';
-import { guilds as storage } from './storage';
-import logger from './logger';
+import { guilds as storage } from '../utils/storage';
+import logger from '../utils/logger';
 
 function buildEmbed (message: string, color?: ColorResolvable): MessageEmbed {
   const embed = new MessageEmbed()
